Single Session Attendance Log

+++ packages/attendance_management/lib/models/attendee.dart
@@ -30,7 +30,7 @@ class AttendeeModel {
   final bool? nonRecoverableError;
   final String? name;
   final int? rowVersion;
-  final int? status;
+  final double? status;
   final String? identifierID;
   final String? individualNumber;
   final AttendeeAdditionalFields? additionalFields;
